Add an environment variable for data config
An environment variable allows you to move the data outside of the
public directory. The default for both the sqlite data and the user
data is still in the public directory, but this allows you to move
it outside of that directory, which you should do if you run this
on a public network.